requirements.txt
asgiref==3.2.3
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
Django==2.0
djangorestframework==3.8.2
google-api-core==1.15.0
google-auth==1.10.0
google-cloud-core==1.1.0
google-cloud-storage==1.24.1
google-resumable-media==0.5.0
googleapis-common-protos==1.6.0
gunicorn==20.0.4
idna==2.7
mysqlclient==1.4.6
netifaces==0.10.7
numpy==1.16.4
oauthlib==3.1.0
pandas==0.23.4
Pillow==5.0.0
pipenv==2018.11.26
protobuf==3.11.2
pyasn1==0.4.8
pyasn1-modules==0.2.7
python-dateutil==2.8.1
pytz==2019.3
requests==2.20.0
requests-oauthlib==1.2.0
rsa==4.0
scikit-learn==0.19.1
scipy==1.4.1
six==1.13.0
sklearn==0.0
sqlparse==0.3.0
urllib3==1.24.3
virtualenv==16.7.2
virtualenv-clone==0.5.3
whitenoise==5.0.1